move remaining panel header components to widgets
