+++ client/src/Tools/_framework/Paths/PublicActivities.tsx
@@ -17,7 +17,7 @@ import {
 import styled from "styled-components";
 
 import { RiEmotionSadLine } from "react-icons/ri";
-import ContentCard from "../../../PanelHeaderComponents/ContentCard";
+import ContentCard from "../../../Widgets/ContentCard";
 import axios from "axios";
 import { createFullName } from "../../../_utils/names";
 import { ContentStructure } from "./ActivityEditor";
