+++ client/src/Widgets/ContentCard.tsx
